Move leftover files and update imports
