+++ apps/client/src/features/auth/components/RequireAuth.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import useAuth from "../../../hooks/useAuth";
+import useAuth from "@/features/auth/hooks/useAuth";
 import { Navigate, Outlet, useLocation } from "react-router-dom";
 
 const RequireAuth = () => {
