+++ apps/client/src/features/auth/hooks/useLogout.js
