+++ apps/client/src/features/products/components/ProductSearchBar.jsx
