+++ apps/client/src/features/vendors/components/VendorSearchBar.jsx
